Change vector coordinates to double
This permits infinite "vector DPI" for GCode-based machines which have do not have the concept of "pixels".
Also it permits postprocessing of paths without huge quantisation noise (required for the upcoming LTT driver)

+++ src/com/t_oster/liblasercut/JobPart.java
@@ -27,12 +27,12 @@ public abstract class JobPart {
 
   public abstract double getDPI();
     
-  public abstract int getMinX();
+  public abstract double getMinX();
 
-  public abstract int getMinY();
+  public abstract double getMinY();
   
-  public abstract int getMaxX();
+  public abstract double getMaxX();
   
-  public abstract int getMaxY();
+  public abstract double getMaxY();
 
 }
