+++ src/com/t_oster/liblasercut/drivers/EpilogCutter.java
@@ -1043,7 +1043,7 @@ abstract class EpilogCutter extends LaserCutter
     return (int) result;
   }
 
-  private double distance(int x, int y, Point p)
+  private double distance(double x, double y, Point p)
   {
     return Math.sqrt(Math.pow(p.x - x, 2) + Math.pow(p.y - y, 2));
   }
